chore: combines everything into one bin crate
There is no intention of disributing a separate library crate for this
tool so there's no need for a separate lib crate
