+++ Cargo.toml
@@ -1,15 +1,30 @@
-[workspace]
-resolver = "3"
-members = ["crates/*"]
-
-[workspace.package]
+[package]
+name = "releasaurus"
+version = "0.1.0"
 edition = "2024"
 
-[workspace.dependencies]
+[features]
+default = []
+forge-tests = []
+
+[dependencies]
+clap = { version = "4.5.45", features = ["derive"]}
 color-eyre = "0.6.5"
+git2 = "0.20.2"
+gitlab = "0.1802.0"
+git-cliff-core = { version = "2.10.0" }
+git-url-parse = "0.4.5"
+glob = "0.3.3"
+indexmap = "2.10.0"
 log = "0.4.27"
+octocrab = "0.44.1"
 paris = "1.5.15"
+regex = "1.11.1"
+reqwest = "0.12.23"
 secrecy = "0.10.3"
 serde = { version = "1.0.219" }
+serde_json = "1.0.143"
 simplelog = { version = "0.12.2" }
+tempfile = "3.20.0"
+tokio = { version = "1.47.1", features = ["rt-multi-thread", "macros"] }
 toml = "0.9.5"
