+++ Justfile
@@ -5,16 +5,9 @@ test *args:
   cargo test {{args}}
 
 test-unit *args:
-    cargo llvm-cov \
-        --lib \
-        --bins \
-        --workspace \
-        --ignore-filename-regex "(_test\.rs$)|(_tests\.rs$)(tests\/)" {{args}}
+    cargo llvm-cov --ignore-filename-regex "(_test\.rs$)|(_tests\.rs$)" {{args}}
 
 test-integration *args:
-    cargo test --test '*' {{args}}
-
-test-all-cov *args:
-  cargo llvm-cov \
-    --workspace \
-    --ignore-filename-regex "(_test\.rs$)|(_tests\.rs$)(tests\/)" {{args}}
+    cargo llvm-cov \
+      --features forge-tests \
+      --ignore-filename-regex "(_test\.rs$)|(_tests\.rs$)" {{args}}
