--- crates/releasaurus-core/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "releasaurus-core"
-version = "0.1.0"
-edition.workspace = true
-
-[dependencies]
-color-eyre = { workspace = true }
-log = { workspace = true }
-secrecy = { workspace = true }
-serde = { workspace = true }
-git-cliff-core = { version = "2.10.0" }
-glob = "0.3.3"
-regex = "1.11.1"
-git2 = "0.20.2"
-indexmap = "2.10.0"
-serde_json = "1.0.143"
-tokio = { version = "1.47.1", features = ["rt-multi-thread", "macros"] }
-gitlab = "0.1802.0"
-reqwest = "0.12.23"
-octocrab = "0.44.1"
-
-[dev-dependencies]
-tempfile = "3.20.0"
