--- crates/releasaurus-core/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod changelog;
-pub mod forge;
-pub mod git;
