--- crates/releasaurus/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "releasaurus"
-version = "0.1.0"
-edition.workspace = true
-
-[dependencies]
-color-eyre = { workspace = true }
-log = { workspace = true }
-paris = { workspace = true }
-secrecy = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-simplelog = {workspace = true, features = ["paris"]}
-toml = { workspace = true}
-clap = { version = "4.5.45", features = ["derive"]}
-git-url-parse = "0.4.5"
-releasaurus-core = { path = "../releasaurus-core", version = "0.1.0" }
-tempfile = "3.20.0"
-
-[package.metadata.cargo-machete]
-ignored = ["paris"]
