--- releasaurus.toml
@@ -1,11 +1,0 @@
-[[package]]
-name = "cli"
-path = "crates/releasaurus"
-
-[[package]]
-name = "core"
-path = "crates/releasaurus-core"
-
-[[package]]
-name = "main"
-path = "."
